Implement a simple Siamese transfroemr and tester
This transformer integrates with our kht and kit extraction functions.
Our five features are the kht and all kit variations

Right now we are only testing on a smaller subset of our fake profile dataset

This is also the first cut of the model and hopefully we can improve performance. The results seem somewhat promising but there are kinks to be worked out with dataset splitting and our evaluation methodology